+++ R/utils.R
@@ -68,7 +68,7 @@ formatFileSpectrumNames <- function(fileIds, spectrumIds,
 }
 
 utils.removePeaks_centroided <- function(int, t) {
-    int[int <= t] <- 0
+    int[int <= t] <- 0L
     int
 }
 
@@ -1295,3 +1295,7 @@ setMethod("reduce", "data.frame",
     parents | children
 }
 
+windowIndices <- function(i, hws, n) {
+    stopifnot(i <= n)
+    max(1L, i - hws):min(n, i + hws)
+}